LayoutTests/fast/canvas/webgl/webgl2-texture-upload-enums-expected.txt

CONSOLE MESSAGE: line 1: TypeError: null is not an object (evaluating 'gl.NO_ERROR')
Test that glTexImage2D and glTexSubImage2D accept new WebGL 2 enum values

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL successfullyParsed should be true (of type boolean). Was undefined (of type undefined).

TEST COMPLETE
